fix: Add non-generated version of GitRef to react code that supports loading a string
